EXAMEN_TP/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
langchain
langchain-groq
streamlit
requests
python-dotenv